+++ tests/src/Functional/InviteReviewTest.php
@@ -41,7 +41,7 @@ class InviteReviewTest extends OrderBrowserTestBase {
   /**
    * {@inheritdoc}
    */
-  protected $defaultTheme = 'classy';
+  protected $defaultTheme = 'starterkit_theme';
 
   /**
    * {@inheritdoc}