+++ tests/src/Functional/ShopAdminTest.php
@@ -27,7 +27,7 @@ class ShopAdminTest extends CommerceBrowserTestBase {
   /**
    * {@inheritdoc}
    */
-  protected $defaultTheme = 'classy';
+  protected $defaultTheme = 'starterkit_theme';
 
   /**
    * {@inheritdoc}